refactor bugfixes
